refactor(feed): extract shared NoteCard component + enrichNotes module (no behavior change)
Pull the per-note rendering and the note-enrichment logic out of the feed into shared
seams, so the planned new locations (profile "latest note", per-user notes page) and
future per-note improvements (reposts, reply indicator, event tags) are implemented
ONCE instead of three times. Behavior-preserving: the /feed renders identically.

Client — ui/src/components/NoteCard.jsx (new):
  The card chrome (avatar + author-name link + timestamp + actions menu + content),
  moved verbatim out of BrainstormFeed's FeedItem, plus the formatTimestamp /
  absoluteTimestamp helpers. Pure presentational: takes an enriched `item`, no data
  logic. BrainstormFeed now renders <NoteCard item={item} /> and drops the inline card
  + the now-unused Link/NoteActionsMenu/NoteContent/formatTimeAgo imports.

Server — src/api/_shared/noteEnrichment.js (new):
  enrichNotes(notes, scanStrfry) — the former enrichAuthors, verbatim — plus
  extractMentionPubkeys + the lookup cap. feedReadPath.js imports it; buildFeed calls
  enrichNotes. The enriched item shape ({ id, pubkey, createdAt, content, author,
  mentions }) is now documented as the contract every read path serves. Display names
  stay POV-independent (self-asserted kind-0); POV-dependent decorations are a noted
  extension point (parameterize by source).

Tests: live-feed-read-path 29/29 (adds E1 — enrichNotes directly), live-feed-feed-page
27/27 (T5/T6/T19/T24/T25 re-pointed to NoteCard + a page-delegates-to-NoteCard check).
Full vite build green; feed browser-verified rendering identically (author link,
relative time + hover, actions menu, entity links, resolved @mention).
